chore(config): sync local configs

+++ .config/fish/config.fish
@@ -1,6 +1,15 @@
-set fish_greeting
+if status is-interactive
+    set fish_greeting
 
-alias cat="bat"
-alias cleanup="pacman -Qdtq | sudo ifne pacman -Rns -"
+    alias cleanup="pacman -Qdtq | sudo ifne pacman -Rns -"
 
-starship init fish | source
+    starship init fish | source
+end
+
+function fix-dns
+    set iface (ip route get 1.1.1.1 | awk 'NR==1{print $5}')
+    set conn (nmcli -g GENERAL.CONNECTION device show $iface)
+    nmcli connection modify $conn ipv4.ignore-auto-dns yes ipv6.ignore-auto-dns yes
+    nmcli device reapply $iface
+    echo "Done: $conn ($iface)"
+end
